+++ Sources/BraintreePayPal/BTPayPalRequest.swift
@@ -104,6 +104,9 @@ import BraintreeCore
 
     /// Optional: The shopper session ID returned from your shopper insights server SDK integration.
     public var shopperSessionID: String?
+    
+    /// Optional: Changes the call-to-action in the PayPal flow. Defaults to `.none`.
+    public var userAction: BTPayPalRequestUserAction
 
     /// Optional: Recurring billing plan type, or charge pattern.
     public var recurringBillingPlanType: BTPayPalRecurringBillingPlanType?
@@ -141,7 +144,8 @@ import BraintreeCore
         enablePayPalAppSwitch: Bool = false,
         shopperSessionID: String? = nil,
         recurringBillingDetails: BTPayPalRecurringBillingDetails? = nil,
-        recurringBillingPlanType: BTPayPalRecurringBillingPlanType? = nil
+        recurringBillingPlanType: BTPayPalRecurringBillingPlanType? = nil,
+        userAction: BTPayPalRequestUserAction = .none
     ) {
         self.hermesPath = hermesPath
         self.paymentType = paymentType
@@ -161,6 +165,7 @@ import BraintreeCore
         self.shopperSessionID = shopperSessionID
         self.recurringBillingDetails = recurringBillingDetails
         self.recurringBillingPlanType = recurringBillingPlanType
+        self.userAction = userAction
     }
 
     // MARK: Public Methods
@@ -187,6 +192,10 @@ import BraintreeCore
         }
 
         experienceProfile["address_override"] = shippingAddressOverride != nil ? !isShippingAddressEditable : false
+        
+        if userAction != .none {
+            experienceProfile["user_action"] = userAction.stringValue
+        }
 
         var parameters: [String: Any] = [:]